+++ powerbox/__init__.py
@@ -1,5 +1,5 @@
 __version__ = "0.5.4"
 
 from .powerbox import PowerBox, LogNormalPowerBox
-from .tools import get_power, angular_average
+from .tools import get_power, angular_average, angular_average_nd
 from .dft import fft, ifft, fftfreq